Use version aliases in composer-dev.json
The engines package depends on the core, with '4.*' version constraint
in its composer.json. This causes conflicts when installing dependencies
because we install the dev-version of core, which doesn't satisfy the
'4.*' constraint. To avoid having to rename engines' composer-dev.json
to composer.json at build time, we alias the latest version of core as
version 4, which will satisfy engines' constraint.